+++ src/backend/MovieChecker.Domain/Models/Dtos/RatingResponse.cs
@@ -3,4 +3,4 @@ namespace MovieChecker.Domain.Models.Dtos;
 /// <summary>
 /// Response for rating operation
 /// </summary>
-public record RatingResponse(int Rating);
+public record RatingResponse(decimal Rating);
